feat: Support pour un path de vault personnalisés

+++ src/main.rs
@@ -4,16 +4,26 @@ mod crypto;
 
 use anyhow::{Context, anyhow};
 use clap::Parser;
+use std::path::PathBuf;
 use rpassword::prompt_password;
 
-use cli::{Cli, Commands};
+use cli::{Cli, Commands, DEFAULT_PATH};
 use json_vault::{Entry, open_vault, write_vault, add_entry, delete_entry, read_entry};
 use crypto::{cipher, decipher};
+
+fn handle_path(path: String) -> PathBuf {
+    if path == DEFAULT_PATH {
+        PathBuf::from(DEFAULT_PATH)
+    } else {
+        PathBuf::from(&path)
+    }
+}
 
 fn main() -> anyhow::Result<()> {
     let cli = Cli::parse();
 
-    let mut entries = open_vault()
+    let path = handle_path(cli.path);
+    let mut entries = open_vault(&path)
         .context("Failed to open vault")?;
 
     match &cli.command {
@@ -47,13 +57,17 @@ fn main() -> anyhow::Result<()> {
                 .context("Failed to create a new entry")?;
             add_entry(&mut entries, String::from(label), entry);
             
-            write_vault(entries)
+            write_vault(path, entries)
                 .context("Failed to save the vault to disk")?;
             println!("Entry '{label}' was successfully saved")
         }
         Commands::Delete { label } => {
+            if !&path.exists() {
+                return Err(anyhow!("Failed to delete entry '{label}', because no vault file was found at path: {path:?}"));
+            }
+
             if delete_entry(&mut entries, String::from(label)) {
-                write_vault(entries)
+                write_vault(path, entries)
                     .context("Failed to save the vault to disk")?;
                 println!("Entry '{label}' was correctly deleted");
             } else {
@@ -61,6 +75,10 @@ fn main() -> anyhow::Result<()> {
             }
         }
         Commands::Read { label } => {
+            if !path.exists() {
+                return Err(anyhow!("Failed to read entry '{label}', because no vault file was found at path: {path:?}"));
+            }
+
             let Some(entry) = read_entry(&entries, &label) else {
                 return Err(anyhow!("Entry '{label}' does not exists, nothing to read"));
             };
